Score:0
Example Accept
Score still 0
review main(),function and class function,than review LinkedBag

+++ main.cpp
@@ -21,7 +21,8 @@ int minute_gap(string,string);
 void to_dsetination(Car,Passenger,string,char*,char);
 int compute_score(int r,int c,int t,int d);
 void search_for_car(Passenger,bool,char*,string);
-
+bool check_CP(Car,Passenger,string);
+bool check_AD(Car,Passenger,string);
 class Entity
 {
 protected:
@@ -34,16 +35,16 @@ protected:
     int lat;
     int tag_cnt;
     int no;
-    string **tag;
+    string tag[20];
 public:
     Entity(string id,bool isOn,int isSer,int lon,int lat,int no,string time_tag):id(id),isOn(isOn),isSer(isSer),lon(lon),lat(lat),no(no),time_tag(time_tag)
     {
         tag_cnt=0;
-        tag=new string*[10000];
     }
     virtual void print()=0;
     void add_tag(string s);
     bool operator==(const Entity&)const;
+    //~Entity();
 };
 
 bool Entity::operator==(const Entity &e)const
@@ -70,6 +71,8 @@ class Car:public Entity
 {
     friend void search_for_car(Passenger,bool,char*,string);
     friend void to_destination(Car,Passenger,string,char*,char);
+    friend bool check_CP(Car,Passenger,string);
+    friend bool check_AD(Car,Passenger,string);
 private:
     int total_score;
     int rate_cnt;
@@ -83,6 +86,7 @@ public:
         total_score=0;
         rate_cnt=0;
     }
+    Car(const Car&);
     void print();
     void get_online(string,int,int,char);
     void change_direction(string,char);
@@ -107,6 +111,8 @@ class Passenger:public Entity
     friend void Car::get_onboard(string,Passenger);
     friend void search_for_car(Passenger,bool,char*,string);
     friend void to_destination(Car,Passenger,string,char*,char);
+    friend bool check_CP(Car,Passenger,string);
+    friend bool check_AD(Car,Passenger,string);
 private:
     int wait;
 public:
@@ -114,6 +120,7 @@ public:
     {
         wait=0;
     }
+    Passenger(const Passenger&);
     void print();
     void get_online(string,int,int,string);
     void get_onboard(string);
@@ -142,6 +149,8 @@ class LinkedBag:public BagInterface<T>
 {
     friend void to_destination(Car,Passenger,string,char*,char);
     friend void search_for_car(Passenger,bool,char*,string);
+    friend bool check_CP(Car,Passenger,string);
+    friend bool check_AD(Car,Passenger,string);
 private:
     Node<T> *head;
     int item_cnt;
@@ -297,11 +306,18 @@ Node<T>* Node<T>::getNext()const
 
 void Entity::add_tag(string s)
 {
-    this->tag[this->tag_cnt]=new string;
-    *this->tag[this->tag_cnt]=s;
+    this->tag[this->tag_cnt]=s;
     this->tag_cnt++;
 }
-
+Car::Car(const Car &c):Entity(c)
+{
+    this->total_score=c.total_score;
+    this->rate_cnt=c.rate_cnt;
+    this->direction=c.direction;
+    this->normal=c.normal;
+    for(int i=0;i<this->tag_cnt;i++)
+        this->tag[i]=c.tag[i];
+}
 void Car::print()
 {
     cout<<this->id<<" ";
@@ -387,21 +403,26 @@ void Car::leave_customer(string time_tag,int lon,int lat,int score,char directio
     this->total_score+=score;
     this->rate_cnt++;
 }
-
+Passenger::Passenger(const Passenger &p):Entity(p)
+{
+    this->wait=p.wait;
+    for(int i=0;i<this->tag_cnt;i++)
+        tag[i]=p.tag[i];
+}
 void Passenger::print()
 {
     cout<<this->id<<" ";
     if(this->isOn==true)
     {
         if(this->isSer==1)
-            cout<<"1 ";
+            cout<<"1";
         else
-            cout<<"2 ";
+            cout<<"2";
     }
     else
-        cout<<"0 ";
+        cout<<"0";
     
-    if(isOn&&isSer)
+    if(isOn&&isSer>0)
         cout<<" "<<this->connect_to;
     cout<<endl;
 }
@@ -431,27 +452,30 @@ void Passenger::get_offline()
     this->isOn=0;
     this->isSer=0;
     this->connect_to="";
+    this->wait=0;
 }
-
 int K,aR,bR,aL,bL,k,h,p,n,car_no=1,passenger_no=1;                                                                    //necessary parameter
-int profit;
+int profit=0,serve_cnt=0;
 LinkedBag<Car>car_registered;
 LinkedBag<Passenger>passenger_registered;
 LinkedBag<Car>car_online;
-LinkedBag<Passenger>passenger_online;
 LinkedBag<Car>car_in_service;
 LinkedBag<Passenger>passenger_in_service;
 int main()
 {
-    ifstream myfile("test.txt");
+    //ifstream myfile("test.txt");
     string useless;
     char statement[1000]={0};
-    myfile>>K>>aR>>bR>>aL>>bL>>k>>h>>p>>n;
-    myfile.get();                                                                              //"\n"
-    getline(myfile,useless);
-    while(!myfile.eof())
+    cin>>K>>aR>>bR>>aL>>bL>>k>>h>>p>>n;
+    cin.get();
+    //myfile>>K>>aR>>bR>>aL>>bL>>k>>h>>p>>n;
+    //myfile.get();                                                                              //"\n"
+    //getline(myfile,useless);
+    getline(cin,useless);
+    while(cin.getline(statement,sizeof(statement)))
+    //while(!myfile.eof())
     {
-        myfile.getline(statement,sizeof(statement));
+        //myfile.getline(statement,sizeof(statement));
         if(strlen(statement)==0)
             break;
         char *statement_time=strtok(statement," ");
@@ -474,7 +498,8 @@ int main()
                 passenger_registered.add(new_p);
                 passenger_no++;
             }
-            cout<<"Aa"<<endl;
+            
+            //cout<<"Aa"<<endl;
         }
         else if(!strcmp(operation,"NC"))
         {
@@ -507,7 +532,7 @@ int main()
                 car_registered.add(new_c);
                 car_no++;
             }
-            cout<<"Aa"<<endl;
+            //cout<<"Aa"<<endl;
         }
         else if(!strcmp(operation,"OC"))
         {
@@ -520,14 +545,13 @@ int main()
             int lon=atoi(strtok(coordinate,","));
             int lat=atoi(strtok(NULL,","));
             Car search_test(id,false,0,0,lon,lat,0,statement_time);
-            if(car_registered.contain(search_test)&&!car_online.contain(search_test))
+            if(car_registered.contain(search_test)&&(!car_online.contain(search_test)&&!car_in_service.contain(search_test)))
             {
-                Node<Car>*to_online=car_registered.getPointerTo(search_test);
-                Car c_get_online=to_online->getItem();
+                Car c_get_online=car_registered.getPointerTo(search_test)->getItem();
                 c_get_online.get_online(statement_time,lon,lat,direction);
                 car_online.add(c_get_online);
             }
-            cout<<"Aa"<<endl;
+            //cout<<"Aa"<<endl;
         }
         else if(!strcmp(operation,"EC"))
         {
@@ -538,13 +562,12 @@ int main()
             Car search_test(id,false,0,0,0,0,0,statement_time);
             if(car_online.contain(search_test))
             {
-                Node<Car>*to_change=car_online.getPointerTo(search_test);
-                Car c_change_direction=to_change->getItem();
+                Car c_change_direction=car_online.getPointerTo(search_test)->getItem();
                 c_change_direction.change_direction(statement_time,direction);
-                car_online.remove(to_change->getItem());
+                car_online.remove(c_change_direction);
                 car_online.add(c_change_direction);
             }
-            cout<<"Aa"<<endl;
+            //cout<<"Aa"<<endl;
         }
         else if(!strcmp(operation,"OP"))
         {
@@ -557,13 +580,12 @@ int main()
             if(!strcmp(p,"L"))
                 normal=false;
             Passenger search_test(id,0,0,0,0,0,statement_time);
-            if(passenger_registered.contain(search_test)&&!passenger_online.contain(search_test))
+            if(passenger_registered.contain(search_test)&&!passenger_in_service.contain(search_test))
             {
-                Node<Passenger>*target_p=passenger_registered.getPointerTo(search_test);
-                Passenger p_online=target_p->getItem();
+                Passenger p_online=passenger_registered.getPointerTo(search_test)->getItem();
                 search_for_car(p_online,normal,coordinate,statement_time);
             }
-            cout<<"Aa"<<endl;
+            //cout<<"Aa"<<endl;
         }
         else if(!strcmp(operation,"CP"))
         {
@@ -577,14 +599,17 @@ int main()
                 Node<Passenger>*catch_p=passenger_in_service.getPointerTo(search_test);
                 Car temp_c=catch_c->getItem();
                 Passenger temp_p=catch_p->getItem();
-                car_in_service.remove(temp_c);
-                passenger_in_service.remove(temp_p);
-                temp_c.get_onboard(statement_time,temp_p);
-                temp_p.get_onboard(statement_time);
-                car_in_service.add(temp_c);
-                passenger_in_service.add(temp_p);
+                if(check_CP(temp_c,temp_p,statement_time))
+                {
+                    car_in_service.remove(temp_c);
+                    passenger_in_service.remove(temp_p);
+                    temp_c.get_onboard(statement_time,temp_p);
+                    temp_p.get_onboard(statement_time);
+                    car_in_service.add(temp_c);
+                    passenger_in_service.add(temp_p);
+                }
             }
-            cout<<"Aa"<<endl;
+            //cout<<"Aa"<<endl;
         }
         else if(!strcmp(operation,"AD"))
         {
@@ -600,9 +625,10 @@ int main()
                 Car destination_c=car_in_service.getPointerTo(search_test)->getItem();
                 Passenger search_test_p(destination_c.customer(),0,0,0,0,0,statement_time);
                 Passenger destination_p=passenger_in_service.getPointerTo(search_test_p)->getItem();
-                to_destination(destination_c,destination_p,statement_time,coordinate,direction);
+                if(check_AD(destination_c,destination_p,statement_time))
+                   to_destination(destination_c,destination_p,statement_time,coordinate,direction);
             }
-            cout<<"Aa"<<endl;
+            //cout<<"Aa"<<endl;
         }
         else if(!strcmp(operation,"LC"))
         {
@@ -617,7 +643,7 @@ int main()
                 car_registered.remove(off_c);
                 car_registered.add(off_c);
             }
-            cout<<"Aa"<<endl;
+            //cout<<"Aa"<<endl;
         }
         else if(!strcmp(operation,"SC"))
         {
@@ -627,7 +653,6 @@ int main()
             if(car_in_service.contain(search_test))
             {
                 Car temp=car_in_service.getPointerTo(search_test)->getItem();
-                temp.change_direction(statement_time);
                 temp.print();
             }
             else if(car_online.contain(search_test))
@@ -639,12 +664,11 @@ int main()
             else if(car_registered.contain(search_test))
             {
                 Car temp=car_registered.getPointerTo(search_test)->getItem();
-                temp.change_direction(statement_time);
                 temp.print();
             }
             else
-                cout<<id<<": no registration"<<endl;
-            cout<<"Aa"<<endl;
+                cout<<id<<": no registration!"<<endl;
+            //cout<<"Aa"<<endl;
         }
         else if(!strcmp(operation,"SP"))
         {
@@ -653,26 +677,23 @@ int main()
             Passenger search_test(id,false,0,0,0,0,statement_time);
             if(passenger_in_service.contain(search_test))
                 passenger_in_service.getPointerTo(search_test)->getItem().print();
-            else if(passenger_online.contain(search_test))
-                passenger_online.getPointerTo(search_test)->getItem().print();
             else if(passenger_registered.contain(search_test))
                 passenger_registered.getPointerTo(search_test)->getItem().print();
             else
-                cout<<id<<": no registration"<<endl;
-            cout<<"Aa"<<endl;
+                cout<<id<<": no registration!"<<endl;
+            //cout<<"Aa"<<endl;
         }
         else if(!strcmp(operation,"SR"))
         {
             cout<<profit<<endl;
-            cout<<"Aa"<<endl;
+            //cout<<"Aa"<<endl;
         }
         else if(!strcmp(operation,"ZZ"))
         {
             car_in_service.clear();
             car_online.clear();
             passenger_in_service.clear();
-            passenger_online.clear();
-            cout<<"Aa"<<endl;
+            //cout<<"Aa"<<endl;
         }
     }
     return 0;
@@ -700,32 +721,33 @@ void to_destination(Car c,Passenger p,string time_tag,char *coordinate,char dire
     int lat=atoi(strtok(NULL,","));
     int duration=minute_gap(c.time_tag,time_tag);
     int distance=abs(lon-c.lon)+abs(lat-c.lat);
+    int t_duration=distance;
     int score=4;
     if(!c.normal)
     {
-        if(distance%1)
-            distance=(distance+1)/2;
+        if(distance%2==1)
+            t_duration=(t_duration+1)/2;
         else
-            distance/=2;
+            t_duration/=2;
     }
     bool free=false;
     if(p.wait>20)
         free=true;
-    if(duration>3*distance)
+    if(duration>3*t_duration)
         free=true;
     
     if(p.wait>20)
         score-=2;
-    else if(10<p.wait&&p.wait<20)
+    else if(10<p.wait&&p.wait<=20)
         score-=1;
-    if(duration>3*distance)
+    if(duration>3*t_duration)
         score-=2;
-    else if(duration>2*distance&&duration<3*distance)
+    else if(duration>2*t_duration&&duration<=3*t_duration)
         score-=1;
     
     for(int i=0;i<c.tag_cnt;i++)
         for(int j=0;j<p.tag_cnt;j++)
-            if(c.tag[i]==c.tag[j])
+            if(!c.tag[i].compare(p.tag[j]))
                 score++;
     
     if(score>5)
@@ -733,16 +755,21 @@ void to_destination(Car c,Passenger p,string time_tag,char *coordinate,char dire
     else if(score<1)
         score=1;
     
-    if(c.normal)
+    if(c.normal&&!free)
         profit+=aR+bR*distance;
-    else
+    else if(!c.normal&&!free)
         profit+=aL+bL*distance;
     c.leave_customer(time_tag,lon,lat,score,direction);
+    car_registered.remove(c);
     car_in_service.remove(c);
     car_online.add(c);
+    c.get_offline(time_tag);
+    car_registered.add(c);
     p.get_offline();
     passenger_in_service.remove(p);
-    passenger_online.add(p);
+    passenger_registered.remove(p);
+    passenger_registered.add(p);
+    serve_cnt++;
 }
 
 int compute_score(int r,int c,int t,int d)
@@ -771,11 +798,14 @@ void search_for_car(Passenger p,bool normal,char *coordinate,string time_tag)
         //============================================================================================
         int distance=abs(p.lon-car_current.lon)+abs(p.lat-car_current.lat);
         if(distance>K)
-            break;
+        {
+            current=current->getNext();
+            continue;
+        }
         int same_tag=0;
         for(int i=0;i<p.tag_cnt;i++)
             for(int j=0;j<car_current.tag_cnt;j++)
-                if(!(*p.tag[i]).compare(*car_current.tag[j]))
+                if(!(p.tag[i]).compare(car_current.tag[j]))
                     same_tag++;
         int a=car_current.total_score;
         int b=car_current.rate_cnt;
@@ -819,6 +849,21 @@ void search_for_car(Passenger p,bool normal,char *coordinate,string time_tag)
     car_in_service.add(to_service);
     passenger_in_service.add(p);
     car_online.remove(to_service);
-    passenger_online.remove(p);
     //============================================================================================
+}
+
+bool check_CP(Car c,Passenger p,string time_tag)
+{
+    int duration=minute_gap(c.time_tag,time_tag);
+    if(c.isSer==2&&p.isSer==1&&duration>0)
+        return true;
+    return false;
+}
+
+bool check_AD(Car c,Passenger p,string time_tag)
+{
+    int duration=minute_gap(c.time_tag,time_tag);
+    if(c.isSer==3&&p.isSer==2&&duration>0)
+        return true;
+    return false;
 }
